requirements.txt
transformers
torch
fastapi
Pillow
uvicorn

